LayoutTests/http/tests/security/contentSecurityPolicy/connect-src-eventsource-blocked-expected.txt

CONSOLE MESSAGE: Refused to load connect from 'http://127.0.0.1:8000/eventsource/resources/simple-event-stream.asis' because of Content-Security-Policy.

Pass
